+++ extensions/iceworks-config-helper/web/src/routes.ts
@@ -1,0 +1,10 @@
+import JSONForm from '@/pages/JSONForm/index';
+
+const routerConfig = [
+  {
+    path: '/',
+    component: JSONForm,
+  },
+];
+
+export default routerConfig;