+++ extensions/iceworks-config-helper/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
 
-## 0.2.2
+## 0.3.0
 
+- feat: add visual editing panel.
+- feat: add rax app.json validation.
 - feat: update browserslist.
+- fix: update ice & rax validation.
 
 ## 0.2.1
 
@@ -10,8 +13,8 @@
 
 ## 0.2.0
 
-- feat: add ICE APP build.json validation
+- feat: add ICE APP build.json validation.
 
 ## 0.1.0
 
-- feat: add build.json validation
+- feat: add build.json validation.